Mobile Track 1: shared HTTP-service model + registry validation
Establishes ONE shared definition of a custom HTTP service and the community
registry manifest, so desktop and mobile don't diverge (and mobile Track 2 can
consume the same core). Pure logic only — platform I/O stays per-platform.

- shared/chat/service-exec.ts: pure request-shaping extracted from
  electron/lib/custom-services.ts (namespacing, parseToolDefinition, coerceArgs,
  buildRequest, filterResponse, sampleArgsFromSchema, serviceToOpenAI) + shared
  types CustomServiceRuntimeConfig / BearerResolver / OpenAIToolDef. electron's
  custom-services.ts now re-exports these and keeps ONLY the impure parts
  (callService/testService/withOAuthBearer/fetchWithTimeout — node fetch +
  keychain resolveSecrets).
- shared/chat/registry-schema.ts: manifest TYPES + Zod schemas + parseManifest
  (entry-level fail-soft). electron's community-registry.ts imports + re-exports
  these and keeps only fetch/ETag/fs-cache. zod 4.4.3 on both platforms.
- shared/chat/service-exec.test.ts (10). Existing custom-services.test.ts (28)
  + community-registry.test.ts (9) still pass via the re-exports.
- mobile changelog v0.1.3: silent-empty-note fix + tappable web results (Track 0)
  + under-the-hood sharing note.

Verify: root 1383 pass/12 skip; type-check:all + lint + compile green; mobile
type-check + lint green. Desktop behavior unchanged (pure refactor).
