+++ electron/lib/community-registry.ts
@@ -17,57 +17,20 @@
 
 import fs from "fs";
 import path from "path";
-import * as z from "zod";
 import { findUserDataDir } from "../runtime/port-discovery";
+import { parseManifest, type CommunityManifest } from "../../shared/chat/registry-schema";
 
-// Local types (main-process side mirrors src/types/index.ts — the two sides are
-// wired by IPC channel strings, not a shared import; see preload.ts convention).
-export interface CommunityManifest {
-  version: number;
-  updatedAt: string;
-  mcpServers: RegistryMcpEntry[];
-  services: RegistryServiceEntry[];
-}
-interface RegistryEntryMeta {
-  id: string;
-  author: string;
-  version: string;
-  category?: string;
-  tags: string[];
-  blurb: string;
-  brandColor?: string;
-  homepage?: string;
-  /** Brand logo, compiled + allowlist-sanitized by cairn-community CI. */
-  iconSvg?: string;
-}
-interface RegistryMcpEntry extends RegistryEntryMeta {
-  definition: {
-    name: string;
-    description?: string;
-    transport: "sse" | "http";
-    baseUrl: string;
-    headers?: Record<string, string>;
-    authMode?: "none" | "oauth";
-    oauthScope?: string;
-    disabledTools?: string[];
-    enabled: boolean;
-  };
-}
-interface RegistryServiceEntry extends RegistryEntryMeta {
-  definition: {
-    name: string;
-    description?: string;
-    apiUrl: string;
-    method: "GET" | "POST" | "PUT" | "DELETE";
-    headers?: Record<string, string>;
-    toolDefinition: string;
-    responseKeys?: string[];
-    apiKeyUrl?: string;
-    authMode?: "none" | "oauth";
-    oauth?: { serverUrl?: string; scope?: string; clientId?: string; authorizationUrl?: string; tokenUrl?: string };
-    enabled: boolean;
-  };
-}
+// Manifest TYPES + Zod validation now live in shared/chat/registry-schema.ts so
+// desktop and mobile validate the catalog identically. Re-export the types so
+// existing electron importers keep resolving them from here.
+export type {
+  CommunityManifest,
+  RegistryMcpEntry,
+  RegistryServiceEntry,
+  RegistryEntryMeta,
+} from "../../shared/chat/registry-schema";
+export { parseManifest } from "../../shared/chat/registry-schema";
+
 export interface RegistryFetchResult {
   manifest: CommunityManifest;
   fromCache: boolean;
@@ -79,93 +42,6 @@ const MANIFEST_URL =
   "https://raw.githubusercontent.com/ddutchie/cairn-community/main/manifest.json";
 const CACHE_FILE = "community-registry.json";
 const FETCH_TIMEOUT_MS = 10_000;
-
-// ── validation (mirrors cairn-community/schema.json) ────────────────────────
-
-const headers = z.record(z.string(), z.string()).optional();
-
-const mcpDefinition = z.object({
-  name: z.string().min(1),
-  description: z.string().optional(),
-  transport: z.enum(["sse", "http"]),
-  baseUrl: z.string().url().startsWith("https://"),
-  headers,
-  authMode: z.enum(["none", "oauth"]).optional(),
-  oauthScope: z.string().optional(),
-  disabledTools: z.array(z.string()).optional(),
-  enabled: z.boolean(),
-});
-
-const oauthConfig = z
-  .object({
-    serverUrl: z.string().url().startsWith("https://").optional(),
-    scope: z.string().optional(),
-    clientId: z.string().optional(),
-    authorizationUrl: z.string().url().startsWith("https://").optional(),
-    tokenUrl: z.string().url().startsWith("https://").optional(),
-  })
-  .optional();
-
-const serviceDefinition = z.object({
-  name: z.string().min(1),
-  description: z.string().optional(),
-  apiUrl: z.string().url().startsWith("https://"),
-  method: z.enum(["GET", "POST", "PUT", "DELETE"]),
-  headers,
-  toolDefinition: z.string().min(1),
-  responseKeys: z.array(z.string()).optional(),
-  apiKeyUrl: z.string().url().startsWith("https://").optional(),
-  authMode: z.enum(["none", "oauth"]).optional(),
-  oauth: oauthConfig,
-  enabled: z.boolean(),
-});
-
-const entryMeta = {
-  id: z.string(),
-  author: z.string(),
-  version: z.string(),
-  category: z.string().optional(),
-  tags: z.array(z.string()),
-  blurb: z.string(),
-  brandColor: z.string().optional(),
-  // Validated as an https URL — it is rendered as an anchor href in the Browse
-  // modal, so an unvalidated string could smuggle a javascript:/data: URI.
-  homepage: z.string().url().startsWith("https://").optional(),
-  // Brand logo, compiled + allowlist-sanitized by cairn-community CI (never raw
-  // user SVG). Rendered inline by ConnectorLogo. Absent → app fallback glyph.
-  iconSvg: z.string().optional(),
-};
-
-const mcpEntry = z.object({ ...entryMeta, definition: mcpDefinition }).passthrough();
-const serviceEntry = z.object({ ...entryMeta, definition: serviceDefinition }).passthrough();
-
-// Manifest-level shape only validates the envelope; entries are validated
-// individually in parseManifest so ONE bad community entry can't blank the
-// whole catalog (the reject-all behaviour of z.array(z.object(...)) would).
-const manifestSchema = z.object({
-  version: z.number(),
-  updatedAt: z.string(),
-  mcpServers: z.array(z.unknown()),
-  services: z.array(z.unknown()),
-});
-
-/** Parse + validate an unknown payload into a CommunityManifest, or throw. */
-export function parseManifest(raw: unknown): CommunityManifest {
-  const m = manifestSchema.parse(raw);
-  return {
-    version: m.version,
-    updatedAt: m.updatedAt,
-    // Drop malformed entries individually rather than failing the whole parse.
-    mcpServers: m.mcpServers.flatMap((e) => {
-      const r = mcpEntry.safeParse(e);
-      return r.success ? [r.data] : [];
-    }),
-    services: m.services.flatMap((e) => {
-      const r = serviceEntry.safeParse(e);
-      return r.success ? [r.data] : [];
-    }),
-  } as CommunityManifest;
-}
 
 // ── cache ───────────────────────────────────────────────────────────────────
