+++ mobile/changelogs/v0.1.3.md
@@ -43,7 +43,8 @@ Polish for the workspace switcher and Knowledge Graph controls.
 - **Folders that differ only in capitalisation are now merged**: Notes filed under `Mobile` and `mobile` (or `Mobile/AI` and `mobile/ai`) were listed as two separate folders. They now collapse into one — matched case-insensitively, keeping the first casing seen as the folder's name. The assistant's folder list dedupes the same way.
 - **Reschedule uses timezone-safe dates**: The project Overview's "due soon" list now compares due dates by local calendar day (matching the rest of the app), so a task due today is never shown a day early/late in some timezones.
 - **Sync screen tidied up**: Removed the duplicate workspace picker from the Sync screen — switch workspaces from the Cairn icon in the Projects header. The Sync screen now focuses purely on sync status, iCloud availability, and conflicts.
-- **The assistant no longer gets stuck on a failed note edit**: When the assistant tried to patch a note but the exact text wasn't found (or the note id was stale), the edit simply reported failure with no reason — so the assistant would blindly repeat the same failing call. It now returns actionable guidance instead: re-read the note and copy its exact current text before retrying, search for the correct note id when it's stale, or use "append to note" to add new content — and is told explicitly not to retry the same call.
+- **The assistant no longer silently loses a note's content on a malformed edit**: On the on-device (Apple) model, if the assistant emitted slightly-malformed JSON for a note write (for example markdown with real line breaks in the middle of the text), the app previously ran the tool with *empty* arguments — silently creating or overwriting a note with no content. The assistant now parses those arguments the same forgiving-but-lossless way the desktop app does (fixing only unambiguous JSON formatting slips, never changing your actual text), and if it still can't parse them it reports the error back to the model to retry instead of writing an empty note.
+- **Tap web results straight from chat**: When the assistant runs a web search (or another external tool that returns a link), its tool chip is now tappable and opens that page in your browser — the same way note/task chips open in-app.
 - **See exactly what differs in a sync conflict**: The Sync Conflicts screen used to show the original note and the conflicted copy as two plain text previews, leaving you to spot the difference by eye. It now shows a unified diff — lines only in the copy are highlighted green (+), lines only in the original red (−), unchanged lines dimmed — with a count of what changed, and says so plainly when the two bodies are actually identical.
 
 - **Open notes and tasks straight from chat**: Chat is now far more reliable at jumping to what it mentions. When the assistant creates or touches a note or task, its little tool chip is now tappable and takes you straight there. `[[Wikilinks]]` in replies now open **tasks** too (not just notes), and both are matched to the exact item at the moment they're shown — so taps no longer occasionally open the wrong item or quietly do nothing.
@@ -80,3 +81,4 @@ Polish for the workspace switcher and Knowledge Graph controls.
 - Moved the Knowledge Graph's WebView document builder (the bundled D3 library plus the force-simulation and radial-sunburst rendering code) out of the 1,000-line graph component into its own file. Purely internal; the graph renders and behaves exactly the same.
 - The "ready to work on" task list now resolves each task's blockers in a single batched database query instead of one lookup per blocker. Same result; fewer queries.
 - The unscheduled-tray drop-zone registration and the sync-conflict diff view were made more render-stable (a memoised diff and a stable zone-ref callback), avoiding needless re-registration / recomputation. No visible change.
+- Moved the tolerant tool-argument parser and the external-link extractor into shared code used by both desktop and mobile, so the assistant handles model JSON identically everywhere. The pure HTTP-service request-shaping (query/body building, argument coercion, response trimming) and the community-registry manifest validation were also moved to shared, laying the groundwork for mobile to consume the same connector catalog as desktop. Purely internal; no visible change.
